Move into ops.rs
