+++ node-graph/gcore/src/lib.rs
@@ -172,5 +172,3 @@ pub type WasmSurfaceHandle = application_io::SurfaceHandle<web_sys::HtmlCanvasEl
 pub type WasmSurfaceHandleFrame = application_io::SurfaceHandleFrame<web_sys::HtmlCanvasElement>;
 
 pub use dyn_any::{WasmNotSend, WasmNotSync};
-
-pub mod expression;
